+++ client-app/router/routes/index.ts
@@ -1,0 +1,2 @@
+export * from "./account";
+export * from "./main";
